+++ examples/demo5.py
@@ -109,7 +109,7 @@ def main(argv):
     #       their r_obs / r_psf = 1.4.
 
     gal_ellip_rms = 0.2             # using "distortion" definition of ellipticity:
-                                    #   e = (a^2-b^2)/(a^2+b^2), where a and be are the 
+                                    #   e = (a^2-b^2)/(a^2+b^2), where a and b are the 
                                     #   semi-major and semi-minor axes, respectively.
     gal_ellip_max = 0.6             # Maximum value of e, to avoid getting near e=1.
     gal_g1 = 0.013                  # Applied shear, using normal shear definition: